fix: fix some clang-tidy errors - remove NOLINT annotations for  (totto-use-fixed-width-types-var), replace it with libc aliases, that expresses the intent better, as this custom rule is not made, to be annoying and you need to annotate every libc call with a NOLInt, but help, and when I replace int usage with the type alias, it marks these cases as intended

+++ src/utils/utf8_helper.h
@@ -13,6 +13,6 @@ typedef struct {
 
 GENERATE_VARIANT_ALL_UTF8_DATA_RESULT()
 
-NODISCARD Utf8DataResult get_utf8_string(const void* data, long size);
+NODISCARD Utf8DataResult get_utf8_string(const void* data, size_t size);
 
 void free_utf8_data(Utf8Data data);
